Add files via upload
In my project i have used a pre-trained model for facial emotion detection through a simple, interactive interface powered by Gradio. 

The system is built using the Transformers library by Hugging Face and leverages the "dima806/facial_emotions_image_detection" model for image classification. 
It use  to predict facial emotions in images by taking a user's image as input and outputting a set of predicted emotions with corresponding confidence scores. 

Gradio is used web-based interface for real-time interaction. we can upload images directly, and the system processes the image using the pre-model prediction pipeline, returning the top five probable emotions. 

The flexibility of the interface allows users to easily experiment with different images and visualize the emotional predictions efficiently, with no need for extensive coding knowledge or setup. 


Gradio is used in this snippet to build a user-friendly web application for facial emotion detection. It allows users to easily upload images and see the model's predictions.

The interface abstracts away the complexities of web development, making it accessible to users who want to interact with machine learning models without extensive coding knowledge. This setup is particularly useful for demonstrations, testing models, or sharing your work with others in an interactive manner.

It is used for expressing the real world data to predict the emotion of the people.